Ordenar la lista de clientas por nombre. Añadir método para migrar campo String a Array<String> en DB

+++ app/src/main/java/ivan/pacheco/cristinalozanobeauty/presentation/client/list/ClientListFragment.kt
@@ -31,6 +31,7 @@ class ClientListFragment: Fragment() {
         super.onCreate(savedInstanceState)
         _binding = ClientListFragmentBinding.inflate(layoutInflater)
         // TODO: migrateFieldClients("treatment", "medication")
+        // TODO: migrateFieldClientsToArray("treatment", "serviceList")
 
         return binding.root
     }
@@ -98,6 +99,23 @@ class ClientListFragment: Fragment() {
                 if (value != null) {
                     val updates = mapOf(
                         newField to value,
+                        oldField to FieldValue.delete()
+                    )
+                    document.reference.update(updates)
+                }
+            }
+        }
+    }
+
+    private fun migrateFieldClientsToArray(oldField: String, newField: String) {
+        val clientsRef = Firestore.db.collection("clients")
+
+        clientsRef.get().addOnSuccessListener { querySnapshot ->
+            for (document in querySnapshot.documents) {
+                val value = document.get(oldField)
+                if (value != null && value.toString().isNotEmpty()) {
+                    val updates = mapOf(
+                        newField to listOf(value),
                         oldField to FieldValue.delete()
                     )
                     document.reference.update(updates)
